Added switch to control aggregation level for reproduction/recruitment
New switch sumR.Level. If not defined, data for estimating R is aggregated at the group/observation level as done previously. If set to "line", data is aggregated at the transect line level (per year) instead. 

Note that this could be extended to also work for aggregation at the location / area level later.