Prevent dropping of transect-years without ptarmigan obs
This re-rolls a change made in e184ea15a01ee5b8c78ed0b7e822b95e90313176. The previous change removed transects in selected years if no ptarmigans were observed in the particular year. This should not be done since it is quite possible to have no ptarmigan observations on a transect in a year in areas with low densities. If these transect-years are dropped, density may be overestimated.
(the toggle excludeNeverObs stll allows excluding transects on which no ptarmigans were EVER recorded)