Changed parameterization for spatial REs
The previous setup ended up with very poor mixing for survival intercepts and spatial sigma in particular. I've discussed this with Matthijs before, and he has found that mixing is affected by the choice of centering REs on the mean vs. on 0. So going to try how it goes if I sample spatial REs explicitly from a normal distribution with mean 0 and add to global intercept.